agents/synthesizer/requirements.txt
websockets>=12.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
neo4j>=5.17.0
py-multibase>=1.0.3
